Make all costs/penalties integral
This results in a slight speed-up, and simplifies the code a bit

+++ hgs_vrptw/include/CommandLine.h
@@ -57,7 +57,8 @@ public:
             else if (std::string(argv[i]) == "-nbGranular")
                 config.nbGranular = static_cast<size_t>(atoi(argv[i + 1]));
             else if (std::string(argv[i]) == "-initialTimeWarpPenalty")
-                config.initialTimeWarpPenalty = atof(argv[i + 1]);
+                config.initialTimeWarpPenalty
+                    = static_cast<size_t>(atoi(argv[i + 1]));
             else if (std::string(argv[i]) == "-penaltyBooster")
                 config.penaltyBooster = atof(argv[i + 1]);
             else if (std::string(argv[i]) == "-minimumPopulationSize")
